pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibe-sync"
version = "0.1.0"
description = "AI context preservation CLI tool for tracking project progress."
requires-python = ">=3.9"
dependencies = [
    "typer",
    "rich",
    "google-generativeai",
    "GitPython",
    "python-dotenv",
    "pydantic",
    "fastmcp",
    "google-cloud-storage",
    "uvicorn",
    "tiktoken",
    "groq",
    "openai"
]

[project.scripts]
vibe-sync = "main:app"

[tool.setuptools]
py-modules = ["main", "vibe_core", "ai_bridge", "cloud", "config", "deploy", "hooks", "models", "server"]

[tool.setuptools.packages.find]
where = ["src"]
